refactor(email_service/send_email): split certificate generation logic into certificate_generator.py (SCRUM-2)
